### 2023-07-08 #5631 @editorialbot
[PRE REVIEW]: lavaanExtra: Convenience Functions for Package *lavaan*
**Submitting author:** @rempsyc (<a href="http://orcid.org/0000-0003-4315-6788">Rémi Thériault</a>)
**Repository:** https://github.com/rempsyc/lavaanExtra
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.6
