### 2023-07-08 #5631 @gkthiruvathukal
@rempsyc Can you please comment on how this is research software? The term "convenience functions" strongly suggests an incremental contribution. There is also no clear statement of need in your paper. I'm also going to open up a query scope on this to get input from the editorial board.
