### 2023-07-08 #5631 @rempsyc
Thank you @gkthiruvathukal. I will try to answer this question below.

I believe the idea of convenience functions does not make an R package ineligible for JOSS, see for example my other paper in review at JOSS, which also has "convenience functions" in the title (#5466). After all, most R packages are of convenience, but as you mention, the key is perhaps the level of incrementation, e.g., of substantial scholarly effort. What can be done in `dplyr` can be done in base R, for example; in this sense it is a package of convenience. Similarly, `lavaanExtra` offers an alternative syntax to `lavaan`, improvements to the generation of figures from other packages, a way to easily fetch relevant statistical information from `lavaan` objects, and output them in APA-formatted Microsoft Word tables, as well as unique features such as detecting indirect paths automatically. In this sense, I believe `lavaanExtra` represents a substantial scholarly effort and sufficient incrementation. Although not the same, it seems like the research community appreciates this software, as it has already been downloaded over 2,000 times from CRAN. Finally, for the statement of need, it is certainly possible for me to revise it if needed.

I hope this answers the question, but please, let me know if you require further clarification. I will wait to get approval before suggesting reviewers.
