### 2023-07-12 #5631 @gkthiruvathukal
@rempsyc Thank you. This is a good response. I'm going to remove the query scope and will now proceed to find an editor. Please continue to be patient as I have a bit of a backlog.
