+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   indirect: indirect.html
   regression: regression.html
   special: special.html
-last_built: 2023-06-25T23:31Z
+last_built: 2023-07-15T21:07Z
 urls:
   reference: https://lavaanExtra.remi-theriault.com/reference
   article: https://lavaanExtra.remi-theriault.com/articles
