### 2023-07-15 #5631 @rempsyc
I have updated the statement of need. For potential reviewers, I would suggest:

Creator of `lavaan` (yrosseel)
Main contributor to `lavaan` (TDJorgensen)
Author of `lavaanPlot` (alishinski)
Author of `tidySEM` (cjvanlissa)

Else from the suggested list:

Jamesuanhoro
Wjakethompson
Smasongarrison
lionel68
