### 2023-07-23 #5631 @arfon
:wave: @TDJorgensen @yrosseel @amesuanhoro @Smasongarrison – would any of you be willing to review this submission for JOSS? The submission under consideration is _lavaanExtra: Convenience Functions for Package *lavaan*_

Many thanks
Arfon
