### 2023-07-24 #5631 @yrosseel
Hello Arfon,

I am afraid I will decline. I am (too) negatively biased when it comes to those tidy/easy/extra packages that add (in my humble opinion) trivial stuff. But that is just me… Perhaps other folks find this great/useful.

Yves.

From: Arfon Smith ***@***.***>
Sent: Sunday, July 23, 2023 9:35 AM
Cc: Yves Rosseel ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: lavaanExtra: Convenience Functions for Package *lavaan* (Issue #5631)

👋 @TDJorgensen <https://github.com/TDJorgensen>  @yrosseel <https://github.com/yrosseel>  @amesuanhoro @smasongarrison <https://github.com/smasongarrison>  – would any of you be willing to review this submission for JOSS? The submission under consideration is lavaanExtra: Convenience Functions for Package lavaan

Many thanks
Arfon
