### 2023-07-24 #5631 @gkthiruvathukal
@yrosseel Thanks for your comments. However, I do note that we put this submission through a query scope, and I found the submitting author's feedback to be satisfactory here.  While I wish authors would use a more glamorous term than *convenience functions*, the author (@rempsyc) did clarify that this package is having some impact and has been cited in at one research paper.

I would like to hope that one of the others who was invited ( @TDJorgensen @jamesuanhoro @smasongarrison) will be willing to edit this paper, which I have determined to be in scope and meeting our scholarly effort criteria.
