### 2023-07-25 #5631 @TDJorgensen
I can review it, but I will be out of the office for the next 3-4 weeks. If your time scale is shorter, perhaps Caspar would be a better choice. `lavaanExtra` sounds like it fits within the scope of his `tidySEM` package, so at first glance I expect @yrosseel to have a point about releasing yet another R package instead of contributing to one that already exists.
