+++ R/lavaan_cor.R
@@ -1,15 +1,15 @@
 #' @title Extract relevant correlation indices from lavaan model
 #'
 #' @description Extract relevant correlation indices from lavaan model through
-#'              `lavaan::parameterEstimates` with `standardized = TRUE`. In this
+#'              [lavaan::parameterEstimates] with `standardized = TRUE`. In this
 #'              case, the correlation coefficient (r) represents the resulting
 #'              `std.all` column.
 #'
 #' @param fit lavaan fit object to extract correlations from
 #' @param nice_table Logical, whether to print the table as a
-#'                   `rempsyc::nice_table` as well as print the
+#'                   [rempsyc::nice_table] as well as print the
 #'                   reference values at the bottom of the table.
-#' @param ... Arguments to be passed to `rempsyc::nice_table`
+#' @param ... Arguments to be passed to [rempsyc::nice_table]
 #' @keywords lavaan structural equation modeling path analysis CFA
 #' @return A dataframe of correlations, including the correlated
 #'         variables, the correlation, and the corresponding p-value.
