+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   indirect: indirect.html
   regression: regression.html
   special: special.html
-last_built: 2023-08-25T21:21Z
+last_built: 2023-08-25T21:24Z
 urls:
   reference: https://lavaanExtra.remi-theriault.com/reference
   article: https://lavaanExtra.remi-theriault.com/articles
