### 2023-08-25 0e3e7fe
adjust if (all(is.na(x$lhs.labels)))
