+++ R/nice_modindices.R
@@ -77,7 +77,7 @@ nice_modindices <- function(fit,
     if (!inherits(labels, "data.frame")) {
       labels <- as.data.frame(as.list(labels))
     }
-    for (i in seq(nrow(x))) {
+    for (i in seq_len(nrow(x))) {
       x[i, "lhs.labels"] <- ifelse(
         x$lhs[i] %in% names(labels),
         labels[which(x$lhs[i] == names(labels))],
