+++ tests/testthat/test-nice_modindices.R
@@ -52,7 +52,7 @@ test_that("nice_modindices auto-labels", {
   skip_if_not_installed("sjlabelled")
   skip_if_not_installed("stringdist")
   x <- HolzingerSwineford1939
-  x <- sjlabelled::set_label(x, label = paste0("I am ", 1:ncol(x), " years old."))
+  x <- sjlabelled::set_label(x, label = paste0("I am ", seq_len(ncol(x)), " years old."))
   fit <- sem(HS.model, data = x)
 
   expect_snapshot(
