### 2023-08-27 392b22e
1st argument of lavaanPlot is not model object
