### 2023-08-27 fc95c76
Joss jou paper (#15)
* 1st argument of lavaanPlot is not model object

* Model names in list
