### 2023-08-27 f2f29dc
nice_lavaanPlot: add doc about sig argument
