### 2023-09-23 #5701 @arfon
@jamesuanhoro – thanks for completing your review. Could you link here any issues you opened on the software repository (https://github.com/rempsyc/lavaanExtra) and write a short summary statement for your review? Thank you!
