### 2023-10-05 rempsyc/lavaanExtra#16 @TDJorgensen
?lavaanExtra in README doesn't work 
For review: https://github.com/openjournals/joss-reviews/issues/5701

In the GitHub repository's instructions to access the package's full list of functions, the following R syntax is given:

```
library(lavaanExtra)
?lavaanExtra
```

But the /man directory does not have a lavaanExtra.Rd file (nor does there appear to be an alias pointing to it).  I'm not sure if @rempsyc intended to have a documentation page for the package as a whole; if not, then the README could instead direct users with this syntax:

```
help(package = "lavaanExtra")
```
