### 2023-10-05 rempsyc/lavaanExtra#17 @TDJorgensen
missing menu item in Example lavaan models
For review: https://github.com/openjournals/joss-reviews/issues/5701

In the floating menu on this page:
https://lavaanextra.remi-theriault.com/articles/example.html
There are links only to the first 6 examples. Example 7 (total effects) seems to be omitted.
