### 2023-10-05 rempsyc/lavaanExtra#18 @TDJorgensen
diagonal elements in lavaan_cov() and lavaan_cor()
For review: https://github.com/openjournals/joss-reviews/issues/5701

The separate `lavaan_cov()` and `lavaan_cor()` functions seem inconsistent with a single `lavaan_reg()` function with the `estimate=` argument to select (un)standardized coefficients.  In fact, the output of `lavaan_cor()` is redundant with any covariances printed by `lavaan_cov(..., estimate = "B")`.  It might be more efficient to make the following lines from `lavaan_cor()` conditional on an argument like `diag=FALSE`, so the variances are ignored `if(!diag)`.

```
not.cor <- which(x$lhs == x$rhs) # Why call this "not.cor" when it means the opposite?
x <- x[-not.cor, ]               # This looks like "not not.cor" when you mean "not cor"
```

Of course, the complication is what symbols to use. It makes sense for `lavaan_cor()` to use _r_ for correlations, but that is also what some of the standardized values from `lavaan_cov()` are (yet they are rather unfortunately labeled "b", despite not being regression slopes). If you wanted to segment out the variances, you could

- reserve `lavaan_cov()` to be only for off-diagonal elements, and it returns what `lavaan_cor()` does when `estimate = "B"` (or perhaps more appropriately for this functionality, when `estimate = "r"`, and `estimate = "sigma"` could indicate unstandardized covariances, triggering `rempsyc::nice_table()` to use a \sigma column header).
- The diagonal elements could be returned by a separate function (e.g., `lavaan_var()`), which could be (unstandardized) variances when `estimate = "sigma"` (with a \sigma^2 column header when `nice_table=TRUE`), but setting `estimate = "r2"` or `"rsq"` could be the flag for standardized values (which are the proportion of total variance, so 1 minus R^2). Or even better, you could calculate 1 minus the standardized variances to actually return R^2 (using that as the table header when `nice_table=TRUE`).
    - Granted, you can extract rows with `op == "r2"` from `parameterEstimates()`, but they don't have _p_ values. The test for variances from `standardizedSolution()` is for the null hypothesis that the residual variance = 0, which would imply the null hypothesis that R^2 = 1.  That is probably not very useful. Instead, it would be trivial for you to calculate a new _p_ values to test whether the standardized variance = 1, which would imply R^2 = 0 (probably a more sensible default test).  That only requires calculating a new Wald _z_ statistic as (est - 1)/_SE_ and using `pnorm()` to get a one-tailed _p_ value (since it can only go in one direction).
