### 2023-10-05 rempsyc/lavaanExtra#19 @TDJorgensen
mistake in nice_fit.Rd citation
For review: openjournals/joss-reviews#5701

/man/nice_fit.Rd includes Schreiber (2017) in the **References**, but the **Description** cites Schreiber et al. (2006).  Note not only the incorrect year but also "et al."
