### 2023-10-05 rempsyc/lavaanExtra#20 @TDJorgensen
hyperlink *.Rd references
For review: openjournals/joss-reviews#5701

Use markdown syntax `<URL>` to create hyperlinks to references in the /man documentations.
