### 2023-10-05 rempsyc/lavaanExtra#22 @TDJorgensen
nonsequitur in workflow tutorial
For review: openjournals/joss-reviews#5701

In https://lavaanextra.remi-theriault.com/articles/fullworkflow.html
the CFA example ends by removing items with the lowest loadings in order to improve fit.  This suggestion is provided without any justification or citation that it is a good idea.  Unreliable items can still be included in a model that accurately depicts their relationships with common factors (including method factors that might explain correlated residuals).  Because this package is apparently targeted to "beginners" (in the Statement of Need), I would recommend excluding any example of a workflow that does not demonstrate best practices.
