### 2023-10-05 rempsyc/lavaanExtra#23 @TDJorgensen
name arguments in write_lavaan() examples
For review: openjournals/joss-reviews#5701

The order of arguments in `write_lavaan()` is rather arbitrary.  That is not a bad thing (I see no reason why any particular order would be favored).  But the vignettes sometimes include examples where the model components are passed without naming the arguments:
https://lavaanextra.remi-theriault.com/articles/fullworkflow.html#sem-example

This might give readers the false impression that they can pass them in any order.
