### 2023-10-05 rempsyc/lavaanExtra#24 @TDJorgensen
lavaan_ind() is actually more general
For review: openjournals/joss-reviews#5701

`lavaan_ind()` only selects rows with `$op == ":="`, which can be any user-defined parameters, not merely indirect effects.  This could have strange consequences for labeling when indirect effects aren't the only type of user-defined parameter in a model (e.g., when slopes differ across groups):

```
fit <- sem(' x1 ~ c(a, b)*x2 ', data = HolzingerSwineford1939,
           group = "school", constraints = 'moderated_slope := a - b')
lavaan_ind(fit)
```

It would be wise to either (a) include a more thorough check for selecting only the rows of parameters that are indirect effects, or (b) to rename the function to reflect it being more general.  There could be an argument to request any specific character (your example can use your arrow: `"\u2192"`) to replace underscores, or even accept a vector of replacements if they are not all the same.
