### 2023-10-05 rempsyc/lavaanExtra#25 @TDJorgensen
write_lavaan(mediation=) is unnecessary
For review: openjournals/joss-reviews#5701

`write_lavaan()` internally handles the `mediation=` and `regression=` arguments identically, except that the `label=` argument is additionally passed to `process_vars()` for `mediation=` components.

```
    if (!is.null(mediation)) {
        mediation <- process_vars(mediation, symbol = "~", label = label,
            title = "[-----------Mediations (named paths)-----------]")
    }
    if (!is.null(regression)) {
        regression <- process_vars(regression, symbol = "~",
            title = "[---------Regressions (Direct effects)---------]")
    }
```

I assume the default is `process_vars()` is `label = FALSE`, but that function does not appear to be from this package (nor did a `??` query return anything).
This is perfectly consistent with the help-page description, but I had to inspect the source code to understand why it was necessary.  It does not appear to be necessary, since one could simply pass the same information from mediation to the regression argument, if the source code merely called `regression <- process_vars(regression, symbol = "~", label = label, ...)` once.  The examples about mediation already show users that the `label=` argument is necessary to define indirect effects.

I can't image, but is there some special scenario that includes regression paths in an SEM that are not part of indirect effects of interest?  Maybe to control for covariates, those effects are not of interest?  But those can always be ignored in a table of results.  To exclude them from indirect effects, the `label=` argument could be a character vector of variable names to include (implying `TRUE` when not empty or `NULL`).
