### 2023-10-05 2c6479c
closes #16 (?lavaanExtra)
