### 2023-10-05 rempsyc/lavaanExtra#16 @rempsyc
Thank you! I can't believe I missed this when I copy-pasted from my other package hehe, happy to have solved this. Also didn't know about `help(package = "lavaanExtra")`, this is quite useful and more direct to reach the index page!
