### 2023-10-05 a9ccb42
closes #17 (missing header in vignette)
