### 2023-10-05 rempsyc/lavaanExtra#17 @rempsyc
Thank you! This has been fixed and users can now use the navigation menu properly.
