### 2023-10-05 8d5201a
closes #19 (Schreiber 2017)
