+++ man/nice_fit.Rd
@@ -28,7 +28,7 @@ unbiased SRMR, AIC, and BIC).
 \description{
 Compares fit from one or several lavaan models. Also
 optionally includes references values. The reference fit values are
-based on Schreiber et al. (2006).
+based on Schreiber (2017).
 }
 \details{
 Note that \code{nice_fit} reports the unbiased SRMR through
