+++ NEWS.md
@@ -1,3 +1,6 @@
+## lavaanExtra 0.1.8.1
+* `?lavaanExtra` now works as expected.
+
 # lavaanExtra 0.1.8
 * CRAN resubmission
 * New function: `nice_modindices`, which outputs modification indices along item labels and a similarity score between the left-hand side and right-hand side items. Useful to assess item redundancy when considering modification indices.
