+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: lavaanExtra
 Title: Convenience Functions for Package 'lavaan'
-Version: 0.1.8
+Version: 0.1.8.1
 Date: 2023-08-25
 Authors@R: 
     person("Rémi", "Thériault", , "remi.theriault@mail.mcgill.ca", role = c("aut", "cre"),
