### 2023-10-05 rempsyc/lavaanExtra#19 @rempsyc
Good catch, thank you! Forgot to update when I made the switch.
