### 2023-10-05 966c669
closes #20 (turn links into hyperlinks)
