+++ man/nice_fit.Rd
@@ -33,8 +33,8 @@ based on Schreiber (2017).
 \details{
 Note that \code{nice_fit} reports the unbiased SRMR through
 \code{\link[lavaan:lavResiduals]{lavaan::lavResiduals()}} because the standard SRMR is upwardly
-biased (https://doi.org/10.1007/s11336-016-9552-7) in a noticeable way for
-smaller samples (thanks to James Uanhoro for this change).
+biased (\url{https://doi.org/10.1007/s11336-016-9552-7}) in a noticeable
+way for smaller samples (thanks to James Uanhoro for this change).
 }
 \examples{
 \dontshow{if (requireNamespace("lavaan", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
@@ -60,7 +60,7 @@ nice_fit(fit)
 \references{
 Schreiber, J. B. (2017). Update to core reporting practices in
 structural equation modeling. \emph{Research in social and administrative pharmacy},
-\emph{13}(3), 634-643. https://doi.org/10.1016/j.sapharm.2016.06.006
+\emph{13}(3), 634-643. \url{https://doi.org/10.1016/j.sapharm.2016.06.006}
 }
 \keyword{CFA}
 \keyword{analysis}
