### 2023-10-05 rempsyc/lavaanExtra#20 @rempsyc
Thank you! This way it's a lot faster to review the references 👍
