### 2023-10-05 6d42a05
closes #23 (use named arguments in `write_lavaan`)
