+++ vignettes/indirect.Rmd
@@ -56,7 +56,9 @@ mediation <- list(speed = M, visual = IV)
 indirect <- list(IV = IV, M = M, DV = DV)
 
 # Write the model, and check it
-model <- write_lavaan(mediation, indirect = indirect, label = TRUE)
+model <- write_lavaan(mediation = mediation, 
+                      indirect = indirect, 
+                      label = TRUE)
 cat(model)
 
 # Fit and plot
@@ -77,7 +79,9 @@ mediation <- list(speed = M, textual = M, visual = IV)
 indirect <- list(IV = IV, M = M, DV = DV)
 
 # Write the model, and check it
-model <- write_lavaan(mediation, indirect = indirect, label = TRUE)
+model <- write_lavaan(mediation = mediation, 
+                      indirect = indirect, 
+                      label = TRUE)
 cat(model)
 
 # Fit and plot
@@ -101,7 +105,9 @@ mediation <- list(speed = M, visual = IV, grade = IV)
 indirect <- list(IV = IV, M = M, DV = DV)
 
 # Write the model, and check it
-model <- write_lavaan(mediation, indirect = indirect, label = TRUE)
+model <- write_lavaan(mediation = mediation, 
+                      indirect = indirect, 
+                      label = TRUE)
 cat(model)
 
 # Fit and plot
@@ -124,7 +130,9 @@ mediation <- list(speed = M, textual = M, visual = IV, grade = IV)
 indirect <- list(IV = IV, M = M, DV = DV)
 
 # Write the model, and check it
-model <- write_lavaan(mediation, indirect = indirect, label = TRUE)
+model <- write_lavaan(mediation = mediation, 
+                      indirect = indirect, 
+                      label = TRUE)
 cat(model)
 
 # Fit and plot
@@ -149,7 +157,9 @@ mediation <- list(speed = M, visual = IV)
 indirect <- list(M = M, DV = DV, IV = IV)
 
 # Write the model, and check it
-model <- write_lavaan(mediation, indirect = indirect, label = TRUE)
+model <- write_lavaan(mediation = mediation, 
+                      indirect = indirect, 
+                      label = TRUE)
 cat(model)
 
 # Fit and plot
@@ -172,7 +182,9 @@ mediation <- list(speed = M, textual = M, visual = IV)
 indirect <- list(M = M, DV = DV, IV = IV)
 
 # Write the model, and check it
-model <- write_lavaan(mediation, indirect = indirect, label = TRUE)
+model <- write_lavaan(mediation = mediation, 
+                      indirect = indirect, 
+                      label = TRUE)
 cat(model)
 
 # Fit and plot
@@ -196,7 +208,9 @@ mediation <- list(speed = M, visual = IV, grade = IV)
 indirect <- list(IV = IV, M = M, DV = DV)
 
 # Write the model, and check it
-model <- write_lavaan(mediation, indirect = indirect, label = TRUE)
+model <- write_lavaan(mediation = mediation, 
+                      indirect = indirect, 
+                      label = TRUE)
 cat(model)
 
 # Fit and plot
@@ -219,7 +233,9 @@ mediation <- list(speed = M, textual = M, visual = IV, grade = IV)
 indirect <- list(IV = IV, M = M, DV = DV)
 
 # Write the model, and check it
-model <- write_lavaan(mediation, indirect = indirect, label = TRUE)
+model <- write_lavaan(mediation = mediation, 
+                      indirect = indirect, 
+                      label = TRUE)
 cat(model)
 
 # Fit and plot
@@ -243,7 +259,9 @@ mediation <- list(speed = M, visual = IV, grade = IV)
 indirect <- list(IV = IV, M = M, DV = DV)
 
 # Write the model, and check it
[…]
