### 2023-10-05 rempsyc/lavaanExtra#23 @rempsyc
Hum, good point. Perhaps it would be interesting to find a way that users can pass the arguments in any order, but the trade-off is that it would require users to always name the passed lists in a specific way. Which I think could be more constraining (if users define those objects elsewhere) than having to name the arguments explicitly. For now, I have simply named the arguments throughout the vignettes and README. Thank you.
