### 2023-10-05 9cbc6b0
fix NEWS file for pkgdown
