+++ NEWS.md
@@ -1,3 +1,6 @@
+# lavaanExtra 0.1.9
+* ✨Incoming✨
+
 ## lavaanExtra 0.1.8.1
 * `?lavaanExtra` now works as expected.
 
