### 2023-10-05 rempsyc/lavaanExtra#22 @rempsyc
Thank you, I agree that we want to demonstrate best practice. I would like to revise this example to reflect that.

In this particular case, I think I simply want to demonstrate the use of the `remove.items` feature, which is not showcased elsewhere. This can be helpful in some scenarios. Say for instance that we are running a big study with lots of scales, and participants can only answer so many questionnaires before they tire. So we want to develop a short scale with only x number of items per dimension. We decide to identify which items to keep by relying on the loadings (so it is not about improving fit anymore).

If I revise the vignette to give this hypothetical scenario to showcase `remove.items`, is this more acceptable or still bad practice?
