### 2023-10-06 rempsyc/lavaanExtra#26 @TDJorgensen
link to issues in vignette
For review: openjournals/joss-reviews#5701

At the bottom of [this vignette](https://lavaanextra.remi-theriault.com/articles/indirect.html#other-scenarios), you state

> If you experience any issues with other scenarios, please open a GitHub issue with your example

Because few users will be familiar with how GitHub works (or the fact that the link to the repository is the little symbol at the top-right of the page), it would be helpful to make that text a link:

```
please open a [GitHub issue](https://github.com/rempsyc/lavaanExtra/issues) with your example
```
