### 2023-10-06 rempsyc/lavaanExtra#27 @TDJorgensen
replace "special cases" with a write_lavaan() vignette
For review: openjournals/joss-reviews#5701

The [Special Cases vignette](https://lavaanextra.remi-theriault.com/articles/special.html) is not very well motivated, and uses only one example, which makes the floating menu on the right seem unnecessary.

You might consider making a write_lavaan.Rmd vignette that describes each `write_lavaan()` argument in a separate section, then simply link to examples of its use in other vignettes/[examples](https://lavaanextra.remi-theriault.com/articles/example.html). For example:

- [Example 4.6](https://lavaanextra.remi-theriault.com/articles/example.html#example-4-6) already illustrates (in)equality constraints
- [Example 5](https://lavaanextra.remi-theriault.com/articles/example.html#example-5-mediation) illustrates user-defined parameters
- The [ML-SEM example](https://lavaanextra.remi-theriault.com/articles/example.html#example-6-multilevel-sem) illustrates the necessity of `write_lavaan(..., custom=)`
- You could reproduce the examples from lavaan's [mean-structures tutorial](https://lavaan.ugent.be/tutorial/means.html) in the examples.Rmd vignette, which illustrate the use of `write_lavaan(..., intercepts=)`, and then link to that in the write_lavaan.Rmd vignette
- There is no `write_lavaan(..., thresholds=)` argument, so currently specifying those would require the `custom=` argument as well.  This would be a good place to show users that.

This would then be a more exhaustive documentation for the `write_lavaan()` function, which you can link to in **See also** on the write_lavaan.Rd help page.
