### 2023-10-06 rempsyc/lavaanExtra#22 @TDJorgensen
> we want to develop a short scale with only x number of items per dimension

I agree that's a good practical example to consider, which would demonstrate the value of `cfa_fit_plot(remove.items=)`.  Perhaps the workflow could tie in estimation of reliability for composites with all items vs. dropping items.  You could use `psych::alpha` to estimate reliability of composites with all items, as well as dropping each item (printed under `Reliability if an item is dropped`), although alpha might change very little because of the tau-equivalence assumption.  You could also loop over items with the `semTools::compRelSEM()` function to calculate "omega" instead.

```
data(cattell, package = "psych") # correlation matrix
mod <- write_lavaan(latent = list(G = rownames(cattell)))
fit <- cfa(mod, sample.cov = cattell, sample.nobs = 277)
compRelSEM(fit) # reliability of composite with 12 indicators
## estimate reliability of all 11-indicator composites
data.frame(Drop = rownames(cattell),
           Omega = sapply(rownames(cattell), USE.NAMES = FALSE,
                          function(i) compRelSEM(fit, omit.indicators = i)))
```

That's just for empirical information about items that would lower reliability if dropped. You would of course want to drop multiple items to shorten a scale, then calculate the composite of that reliability.  But you can pass multiple indicators to the `compRelSEM(omit.indicators =)` argument, if you want to estimate it from the larger model first.  That always opens a Pandora's box of "which model is right?", which is an important assumption for factor-model reliability estimation.
