### 2023-10-06 rempsyc/lavaanExtra#28 @TDJorgensen
lavaan_reg() defaults
For review: openjournals/joss-reviews#5701

`lavaan_reg()` only allows for reporting unstandardized or standardized coefficients, the latter being reported with NHSTs based on delta-method *SE*s.  This does not conform to the recommended practice of reporting unstandardized coefficients with their NHSTs (*SE*s, CIs, Wald *z* tests; [McDonald & Ho, 2002](https://doi.org/10.1037//1082-989X.7.1.64), p. 76).

APA recommends reporting a standardized measure of effect size with any NHST, and the standardized coefficient satisfies this without any need to report a redundant NHST for the standardized coefficient (which typically tests the same null hypothesis using different point and *SE* estimates).  A more appropriate default table would reporting the `parameterEstimates()` output with the `std.all` column only used as an effect size (which is why it is reported that way in lavaan's `summary()` method).

However, [Kelley & Preacher (2012)](https://doi.org/10.1037/A0028086) note that standardized estimates might also estimate a specific parameter of interest. They have sampling distributions, so reporting CIs for standardized coefficients is a good idea for the same reason it is good to provide them for unstandardized estimates.

One idea might be to have a `test=` argument in addition to `estimate=`, to indicate which NHST (and CI) to include.  The default might be `lavaan_reg(..., estimate=c("b","B"), test = "b")` to request both estimates but only test the unstandardized.  But it could get complicated figuring out ideal layouts for the 9 different combinations of either/or/both for `estimate=` and `test=`.

At which point the question becomes whether to keep the function simple (expecting it to satisfy all situations) or make it more flexible with options/arguments that give users enough control to publish according to their preferences rather than a package designer's.
