### 2023-10-06 rempsyc/lavaanExtra#29 @TDJorgensen
intuition in Statement of Need
For review: openjournals/joss-reviews#5701

Regarding the paper:

The Statement of Need opens by stating lavaan's popularity, yet claims that the lavaan syntax is not very intuitive for beginners, which contradicts the stated purpose of the lavaan package numerous times by Rosseel (2012, p. 3), whose conclusion reads: "One of the main attractions of lavaan is its intuitive and easy-to-use model syntax" (p. 33). The degree to which a user finds it intuitive is clearly subjective (e.g., that varies widely across my SEM students every year), but this statement seems altogether unnecessary given that the second paragraph does not describe the modular alternative syntax as being more intuitive (which I don't think it is). The value of the alternative syntax approach seems to be the flexibility its modularity (for updating or appending models), which is not necessarily more intuitive.
