### 2023-10-06 6f1d03f
closes #26 (add link to open issue)
