### 2023-10-06 rempsyc/lavaanExtra#30 @TDJorgensen
clarify target audience
For review: openjournals/joss-reviews#5701

Regarding the paper:

I don't think the Statement of Need "clearly identifies ... who the target audience is", which is on the checklist.  It refers to "beginners" and "advanced users", without clarifying whether these are beginners or advanced users of SEM, of R, or of lavaan in particular.  After going through all of the examples, I can't find a reason to recommend it to my students who are learning SEM.  The only reason I would expect to use it myself (or recommend it to colleagues) would be for the nice looking tables, since the figures are simply wrappers around functions already available in other packages.
