### 2023-10-06 rempsyc/lavaanExtra#31 @TDJorgensen
claims about modular syntax approach
For review: openjournals/joss-reviews#5701

Regarding the paper:

The modular syntax is a cool idea, and I have used the same approach (applying `paste()` to lists) when coding simulations.  But it does not do what the paper claims on p. 1 (lines 35-36):

> the user only needs to change it once, at the appropriate location, and it will update future occurrences automatically since it relies on reusable components

This makes it sound like the model is a dynamically changing object, but the vignettes demonstrate that the model needs to be written again with different arguments (e.g., defining a new `regression` object [here](https://lavaanextra.remi-theriault.com/articles/fullworkflow.html#path-analysis-model) before running `write_lavaan()` again).  The `write_lavaan()` function's main purpose is to `paste()` a lavaan operator between a list's names and its contents, with some .  So users have to remember the specific arguments to pass different lists to, rather than remembering the short list of operators listed on the `?lavaan::model.syntax` help page, the [JSS paper](https://doi.org/10.18637/jss.v048.i02), the [website's tutorials](https://lavaan.ugent.be/tutorial/), etc.  The [examples](https://lavaanextra.remi-theriault.com/articles/example.html) make it pretty clear that every other aspect of model syntax needs to be written out (however long the right-hand side needs to be), except for the operator.  So there is still plenty of opportunity for making typos in those lists.

Reusing script components (after typos are fixed) is indeed good practice, but modular syntax does not require using `paste()` the way `write_lavaan()` does.  The advantage of `paste()` for quickly specifying a list of indicators (lines 37-38) is also not unique to lavaanExtra's vector-based approach.  This has always been possible:
```
mod <- paste0("g =~ x", 1:9)
fit <- cfa(mod, data = HolzingerSwineford1939)
```

When teaching SEM to beginners, I sometimes use a modular approach by storing each model component as a single character string (which are more legible than the lists passed to `write_lavaan`) and concatenate the different components when fitting a model to data.  For example, measurement and structural components can be stored in separate strings (e.g., `mod.meas <- "..."` and `mod.reg <- "..."`), then calling the "full SEM" with `sem(c(mod.meas, mod.reg), data = ...)`.  Or when I teach about invariance, I will specify in separate strings:

1. the labeled loadings (sufficient for configural invariance)
2. equality constraints on those labels (concatentated for metric invariance)
3. equality-constrained intercepts (additionally concatenated for scalar invariance)
4. equality-constrained residual variances (additionally concatenated for strict invariance)

These are all "reusable components" (like 36) that obey the DRY principle (line 29).  The `write_lavaan()` function isn't really about modularity.  It is about not writing lavaan syntax yourself, and instead specifying lists of righthand sides, named by lefthand sides.  I think that is a matter of taste, and others may find it as useful as you do.  But I don't see how there is any room to claim that `write_lavaan()` is necessary to achieve the advantages of modularity listed on the first couple pages.  The main advantage of `write_lavaan()` that I can identify is the automated specification of indirect effects, at least in the case of simple and parallel (but not serial) mediation models.
