### 2023-10-06 rempsyc/lavaanExtra#32 @TDJorgensen
demonstrating the value of the plot wrappers
For review: openjournals/joss-reviews#5701

Regarding the paper:

Beginning on p. 4, there is a comparison of the default plots from `lavaanPlot()` and `nice_lavaanPlot()`, but the source code look like it only overrides a few defaults.  There is nothing inherently better or worse about a "naked" path diagram.  It would be helpful for users to see how much time `nice_lavaanPlot()` saves them by providing the `lavaanPlot()` syntax that would yield the same figure as provided by `nice_lavaanPlot()` by default.

Likewise for showing the `tidySEM::graph_sem()` that yields the default `nice_tidySEM()` plot.  Judging from the source code, I imagine `nice_tidySEM()` saves quite some space for users!
