### 2023-10-06 rempsyc/lavaanExtra#33 @TDJorgensen
other output formats for tables
For review: openjournals/joss-reviews#5701

I love the table formats!  I would like to have them accessible for my LaTeX documents, since I no longer use Word (for the same DRY and modularity principles your paper mentions).  The [`semTable`](https://cran.r-project.org/package=semTable) package provides only LaTeX and HTML output (or saved as a CSV file) for tables of lavaan output.  The `knitr::kable()` function does provide LaTeX or Markdown formats (the latter can also be rendered as Word), but I have to specify all the table cells myself (rather than your nice row/column headers).

So this is a feature request, perhaps not doable in the time it takes for a review.   It looks like the [`flextable`](https://cran.r-project.org/package=flextable) package does not output LaTeX, but does output Markdown.  There might be an easy way to at least make that available.  RMarkdown could output a LaTeX file to capture it from there...
