### 2023-10-06 f073922
closes #27 (replace "special cases" vignette with write_lavaan)
