### 2023-10-06 rempsyc/lavaanExtra#27 @rempsyc
Thanks. Done. I have deleted the Special cases vignette and replaced it with one called "How to use `write_lavaan()`" linking to other vignettes as you suggested. I could not find a good example for showcasing thresholds with the "|" operator online. Do you know of any simple example?
