+++ pkgdown.yml
@@ -6,8 +6,8 @@ articles:
   fullworkflow: fullworkflow.html
   indirect: indirect.html
   regression: regression.html
-  special: special.html
-last_built: 2023-10-06T14:09Z
+  write_lavaan: write_lavaan.html
+last_built: 2023-10-06T18:54Z
 urls:
   reference: https://lavaanExtra.remi-theriault.com/reference
   article: https://lavaanExtra.remi-theriault.com/articles
