+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   indirect: indirect.html
   regression: regression.html
   write_lavaan: write_lavaan.html
-last_built: 2023-10-06T18:54Z
+last_built: 2023-10-06T19:02Z
 urls:
   reference: https://lavaanExtra.remi-theriault.com/reference
   article: https://lavaanExtra.remi-theriault.com/articles
