### 2023-10-06 rempsyc/lavaanExtra#33 @rempsyc
Thank you, I appreciate it!

In theory, flextable can export to PDF, e.g., with `output: pdf_document` in an `rmarkdown` document. I've never worked with LaTeX myself (still stuck with Word unfortunately for track-change and collaborating with coauthors who don't use R and don't program) so I don't know how that translates to your exact format. The tables are, however, not as pretty (but still relatively intact):

![pdf_output](https://github.com/rempsyc/lavaanExtra/assets/13123390/91555e83-d021-4d44-a4e8-2a49ceb1f627)

Furthermore, when trying `nice_fit()`, I am getting an error that some symbols are not supported in LaTeX:

```r
! LaTeX Error: Unicode character χ (U+03C7)
               not set up for use with LaTeX.
```

However, [one source](https://bookdown.org/yihui/rmarkdown-cookbook/latex-unicode.html) explains
> If you run into an error like this, it probably means that you are using the default LaTeX engine, pdflatex, to render the (intermediate .tex) document to PDF, which was unable to process certain Unicode characters in your document. If that is the case, you may switch to xelatex or lualatex, e.g.,
> ```
> output:
>   pdf_document:
>     latex_engine: xelatex
> ```

Doing so I was able to get the fit table too.

![fit_table](https://github.com/rempsyc/lavaanExtra/assets/13123390/115a2f7f-05e7-4f13-927a-ad81e18fc0f3)

Resulting PDF: [test_pdf.pdf](https://github.com/rempsyc/lavaanExtra/files/12835013/test_pdf.pdf)

I am happy to learn more about your LaTeX setup to see how we could make it compatible if it does not work already as in the `rmarkdown` to PDF example.
