### 2023-10-06 c54a320
closes #32 (add comparisons to other plotting packages)
