+++ paper/paper.Rmd
@@ -16,11 +16,11 @@ affiliations:
 date: "`r Sys.Date()`"
 bibliography: paper.bib
 output:
-  # rticles::joss_article
-  #   fig_crop: yes
-  md_document:
-    preserve_yaml: TRUE
-    variant: "markdown_strict"
+  rticles::joss_article:
+    fig_crop: yes
+  # md_document:
+  #   preserve_yaml: TRUE
+  #   variant: "markdown_strict"
 journal: JOSS
 link-citations: yes
 csl: apa.csl
@@ -209,8 +209,8 @@ lavaan_ind(fit.sem, nice_table = TRUE)
 ## Figures
 
 There are several packages designed to plot SEM models, but few that
-people consider satisfying or sufficiently good for publication. There
-are two packages that stand out however, {lavaanPlot}
+people consider satisfying or sufficiently good for publication by default.
+There are two packages that stand out however, {lavaanPlot}
 [@lavaanPlotPackage] and {tidySEM} [@tidySEMPackage]. Yet, even for
 those excellent packages, most people do not view them as
 publication-ready or at least optimized in the best possible way.
@@ -223,6 +223,21 @@ side-by-side for demonstration purposes.
 lavaanPlot::lavaanPlot(model = fit.sem)
 
 nice_lavaanPlot(fit.sem)
+```
+
+For reference, `nice_lavaaPlot()` is a simple wrapper around `lavaanPlot::lavaanPlot()` and an identical figure can be obtained using only `lavaanPlot` with the following code:
+
+```{r lavaanPlot, eval=FALSE}
+lavaanPlot::lavaanPlot(
+    model = fit.sem,
+    node_options = list(shape = "box", fontname = "Helvetica"),
+    coefs = TRUE, 
+    stand = TRUE,
+    stars = c("regress", "latent", "covs"), 
+    graph_options = c(rankdir = "LR"),
+    sig = .05
+  )
+
 ```
 
 As these figures demonstrate, `nice_lavaanPlot()` has several elements
@@ -259,6 +274,28 @@ feeding it the `indirect` object that we created earlier.
 
 ```{r nice_tidySEM, fig.width=8, fig.height=4, message=FALSE}
 nice_tidySEM(fit.sem, layout = indirect)
+```
+
+For reference, below I provide the code necessary to reproduce this figure using the `tidySEM` package only.
+
+```{r tidySEM, fig.width=8, fig.height=4, message=FALSE, eval=FALSE}
+library(tidySEM)
+
+mylayout <- data.frame(
+  IV = c("grade", "ageyr"),
+  M = c("", "visual"),
+  DV = c("textual", "speed")
+)
+p <- prepare_graph(fit.sem, layout = mylayout)
+p <- hide_var(p)
+x <- p$edges$est_sig_std
+x <- sub("^0", "", x)
+x <- sub("^-0", "-", x)
+p$edges$label <- x
+p$edges$linetype <- 1
+p$edges$arrow <- ifelse(p$edges$arrow == "none", "both", p$edges$arrow)
+plot(p)
+
 ```
 
 For the time being, `nice_tidySEM` only supports this three-level
@@ -300,6 +337,34 @@ The resulting figure can be saved using `ggplot2::ggsave()`
 
 ```{r, eval=FALSE}
 ggplot2::ggsave("my_semPlot.pdf", width = 8, height = 6)
+```
+
+For reference, below I provide the code necessary to reproduce this figure using the `tidySEM` package only.
+
[…]
