### 2023-10-06 rempsyc/lavaanExtra#32 @rempsyc
Thank you! Done. Although it lengthens the paper, I agree that it is both transparent whilst more explicitly showing the utility. In the case of tidySEM for example, it is not only about the number of lines of code, but also knowing which functions to use to generate the desired output. There are also other arguments that are FALSE by default but that also save time should they be used relative to the original functions, but I think that would be too much for the paper (JOSS papers are supposed to be short and mine is already longer than usual).
