### 2023-10-06 01316ef
closes #29 and closes #30 and closes #31
