+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   indirect: indirect.html
   regression: regression.html
   write_lavaan: write_lavaan.html
-last_built: 2023-10-06T21:26Z
+last_built: 2023-10-06T22:30Z
 urls:
   reference: https://lavaanExtra.remi-theriault.com/reference
   article: https://lavaanExtra.remi-theriault.com/articles
