### 2023-10-06 rempsyc/lavaanExtra#29 @rempsyc
Thank you for bringing this point, I have removed the sentence,
> but this syntax is neither very intuitive for beginners nor very modular for more advanced users

And emphasized the package flexibility and modularity instead.
