### 2023-10-06 rempsyc/lavaanExtra#30 @rempsyc
Thank you, I have revised the first paragraph as such:

> {lavaan} [@lavaanPackage] is a very popular R package for structural
> equation modeling (SEM). The package relies on specific operators to define
> latent variables, regressions, covariances, indirect effects, and so on.
> However, some individuals (e.g,. beginners to R and {lavaan}) may prefer
> not having to specify the operators themselves, or would like to see some
> steps automatized, such as defining indirect effects. Furthermore, for
> researchers, it can be relatively difficult to extract relevant statistical
> outputs in the form of tables and figures that are suitable for scientific
> publication.

And I am open to further suggestions for improvement.
