### 2023-10-06 rempsyc/lavaanExtra#31 @rempsyc
Thank you for this in-depth analysis and explanations, which I highly appreciate. I would like to adapt the paper to better reflect the points you mention.

First, I have deleted the sentence from lines 35-36 about automatically updating future occurrences and replaced it with:
> With `write_lavaan()`, users only need to define the reusable component the first
time, or until they need to change that component again.

Second, in the new version, I have also tried to emphasize that `lavaan` is also compatible with a modular approach, like you have explained it in this issue. Therefore, I have revised the second paragraph of the Statement of Need as follows:
>  {lavaanExtra} does mainly two things to address these issues. First, it offers
an alternative, code-efficient flexible modular syntax that allows automatizing
certain steps, such as defining indirect effects in certain scenarios or the
desired structure of a SEM model to be plotted (however, note that {lavaan} is
also compatible with a modular approach).

And I am open to further suggestions for improvement for this as well.

Later in the **Usage** section (4th paragraph), when discussing modularity and the ability to use paste, I have also added the following sentence:
> However, note that reusable components through functions is also compatible with {lavaan}.
