### 2023-10-07 rempsyc/lavaanExtra#27 @TDJorgensen
Wow, I forgot how sparse lavaan's documentation is for thresholds, even on the `?model.syntax` help page.  Because they are typically "nuisance parameters" or at least not of any substantive interest (like residual variances), they are rarely included directly in model syntax.  I have some simple 2-variable examples here using simulated data, just to demonstrate how threshold models work:

https://tdjorgensen.github.io/TDJorgensen/CatMeasEq/Jorgensen.CatMeasEq.pdf

But a more substantively interesting example will probably always be more complex (typically, imposing invariance constraints).  Here is an example for longitudinal invariance:

https://thechangelab.stanford.edu/tutorials/growth-modeling/modeling-change-latent-vars-ordinal-indicators-lavaan/

Given that it builds up model components by adding equality constraints, it might make a good widely-applicable example to show how your approach would work for evaluating invariance assumptions (as I mentioned in #31).
