### 2023-10-07 rempsyc/lavaanExtra#33 @TDJorgensen
> still stuck with Word unfortunately for track-change and collaborating with coauthors who don't use R and don't program

I won't try to convert you :-) but [Overleaf](https://www.overleaf.com/) provides a great environment for collaborating on a paper in LaTeX, including editing in either the plain-text syntax editor or making edits to a rich-text editor (wysiwyg like Word, or maybe wysiwym like [LyX](https://www.lyx.org/)).  Accounts are free, but the "pro" account also has a track-changes feature (and GitHub integration).

> some symbols are not supported in LaTeX

Yes, math symbols are typically added to LaTeX in math mode:

```
... model was rejected, $\chi^2 (36) = 116.26$, $p < .001$,
and approximate fit was mediocre, RMSEA = 0.90, 90\% CI [0.07, 0.10], ...
```

When I use `knitr::kable()`, I create a matrix/table with character strings (e.g., `"$chi^2$"`) to be interpreted as LaTeX.

But the xelatex compilation is also a solution to continue including unicode, as is exporting to PDF which is easily included (as an image rather than a text-based table) in a LaTeX document.

> I am happy to learn more about your LaTeX setup to see how we could make it compatible if it does not work already as in the `rmarkdown` to PDF example

Honestly, I think if the `flextable` package already supports this, all you should need to do at this point is mention that in the paper and vignettes when you mention exporting to *.docx.
