### 2023-10-07 rempsyc/lavaanExtra#27 @rempsyc
Oh, great, thanks. Do I have your permission to reuse your examples and linking to those files as the source?
