### 2023-10-07 rempsyc/lavaanExtra#27 @TDJorgensen
Yes
