### 2023-10-07 d5f7246
add threshold argument and example (#27) + add tests
