+++ tests/testthat/test-write_lavaan.R
@@ -26,143 +26,49 @@ constraint.equal <- list(b1 = "(b2 + b3)^2")
 constraint.smaller <- list(b1 = "exp(b2 + b3)")
 constraint.larger <- list(b1 = "exp(b2 + b3)")
 custom <- "y1 + y2 ~ f1 + f2 + x1 + x2"
-
+threshold <- list(y2w1 = "thr1*t1", y2w2 = "thr2*t1")
 
 #   ____________________________________________________________________________
 #   Tests                                                                   ####
 
 test_that("write_lavaan using latent", {
-  HS.model <- write_lavaan(latent = latent)
-  expect_equal(
-    HS.model,
-    c(
-      "##################################################
-# [-----Latent variables (measurement model)-----]
-
-visual =~ x1 + x2 + x3
-textual =~ x4 + x5 + x6
-speed =~ x7 + x8 + x9\n\n"
-    )
-  )
+  expect_snapshot(cat(write_lavaan(latent = latent)))
 })
 
 test_that("write_lavaan using regression", {
-  HS.model <- write_lavaan(regression = regression)
-  expect_equal(
-    HS.model,
-    c(
-      "##################################################
-# [---------Regressions (Direct effects)---------]
-
-speed ~ ageyr + grade
-textual ~ ageyr + grade\n\n"
-    )
-  )
+  expect_snapshot(cat(write_lavaan(regression = regression)))
 })
 
 test_that("write_lavaan using covariance", {
-  HS.model <- write_lavaan(covariance = covariance)
-  expect_equal(
-    HS.model,
-    c(
-      "##################################################
-# [------------------Covariances-----------------]
-
-speed ~~ textual
-ageyr ~~ grade\n\n"
-    )
-  )
+  expect_snapshot(cat(write_lavaan(covariance = covariance)))
 })
 
 test_that("write_lavaan using mediation", {
-  HS.model <- write_lavaan(mediation = mediation)
-  expect_equal(
-    HS.model,
-    c(
-      "##################################################
-# [-----------Mediations (named paths)-----------]
-
-speed ~ visual
-textual ~ visual
-visual ~ ageyr + grade\n\n"
-    )
-  )
+  expect_snapshot(cat(write_lavaan(mediation = mediation)))
 })
 
 test_that("write_lavaan using mediation with labels", {
-  HS.model <- write_lavaan(mediation = mediation, label = TRUE)
-  expect_equal(
-    HS.model,
-    c(
-      "##################################################
-# [-----------Mediations (named paths)-----------]
-
-speed ~ visual_speed*visual
-textual ~ visual_textual*visual
-visual ~ ageyr_visual*ageyr + grade_visual*grade\n\n"
-    )
-  )
+  expect_snapshot(cat(write_lavaan(mediation = mediation, label = TRUE)))
 })
 
 test_that("write_lavaan using mediation with letters", {
-  HS.model <- write_lavaan(
+  expect_snapshot(cat(write_lavaan(
     mediation = mediation,
-    label = TRUE, use.letters = TRUE
-  )
-  expect_equal(
-    HS.model,
-    c(
-      "##################################################
-# [-----------Mediations (named paths)-----------]
-
-speed ~ a_speed*visual
-textual ~ a_textual*visual
-visual ~ a_visual*ageyr + b_visual*grade\n\n"
-    )
-  )
+    label = TRUE,
+    use.letters = TRUE)))
 })
 
 test_that("write_lavaan using mediation with letters and indirect", {
-  HS.model <- write_lavaan(
-    mediation = mediation, indirect = indirect,
-    label = TRUE, use.letters = TRUE
-  )
-  expect_equal(
-    HS.model,
-    c(
-      "##################################################
-# [-----------Mediations (named paths)-----------]
-
[…]
