### 2023-10-07 f6ca426
Add Example 10 (longitudinal invariance) (#27)
