### 2023-10-07 beb1d4e
closes #33 (mention PDF export option)
