+++ README.md
@@ -45,7 +45,7 @@ install.packages("lavaanExtra", repos = c(
 Or from GitHub, for the very latest version:
 
 ``` r
-# If package `remotes` isn't already installed, install it with `install.packages("remotes")`
+# If not already installed, install package `remotes` with `install.packages("remotes")`
 remotes::install_github("rempsyc/lavaanExtra")
 ```
 
@@ -335,7 +335,7 @@ fit.sem <- sem(model, data = HolzingerSwineford1939)
 ```
 
 ``` r
-# Get regression parameters only and make it pretty with the `rempsyc::nice_table` integration
+# Get regression parameters and make pretty with `rempsyc::nice_table`
 lavaan_reg(fit.sem, nice_table = TRUE, highlight = TRUE)
 ```
 
@@ -359,8 +359,13 @@ fit_table
 ``` r
 # Save fit table to Word!
 flextable::save_as_docx(fit_table, path = "fit_table.docx")
+# Note that it will also render to PDF in an `rmarkdown` document 
+# with `output: pdf_document`, but using `latex_engine: xelatex` 
+# is necessary when including Unicode symbols in tables like with 
+# the `nice_fit()` function.
 
-# Let's get the indirect effects only and make it pretty with the `rempsyc::nice_table` integration
+# Let's get the indirect effects only and make it pretty 
+# with the `rempsyc::nice_table` integration
 lavaan_ind(fit.sem, nice_table = TRUE)
 ```
 
