+++ vignettes/fullworkflow.Rmd
@@ -90,6 +90,10 @@ fit_table <- nice_fit(lst(fit.cfa, fit.cfa2), nice_table = TRUE)
 flextable::save_as_docx(fit_table, path = "fit_table.docx")
 ```
 
+Note that it will also render to PDF in an `rmarkdown` document with `output: pdf_document`, 
+but using `latex_engine: xelatex` is necessary when including Unicode symbols
+in tables like with the `nice_fit()` function.
+
 ```{r, echo=FALSE}
 unlink("fit_table.docx")
 ```
