### 2023-10-07 rempsyc/lavaanExtra#33 @rempsyc
> I won't try to convert you :-) but [Overleaf](https://www.overleaf.com/) provides a great environment for collaborating on a paper in LaTeX

Hum, it definitely is in my long-term goals to move to fully open-source software, so Overleaf does look like a great starting point for writing! Thank you for the suggestion. Actually I just created my account and started a test project. I love that it recompiles automatically when saving! It is in this way more similar to traditional "what-you-see-is-what-you-get" text processing software. Sadly it seems the free version is only limited to one collaborator and that my university does not have a license :(

Also, I added the following sentence in the paper, README, and vignette:

> It will also render to PDF in an `rmarkdown` document with `output: pdf_document`,
but using `latex_engine: xelatex` is necessary when including unicode symbols
in tables like with the `nice_fit()` function.
