### 2023-10-07 rempsyc/lavaanExtra#21 @rempsyc
This is very useful, thank you. It probably is the most critical for publishing the paper given that it includes tables with this footnote, and I would not want to crystalize bad practice in a paper such as this one.

First of all, may I have your permission to include the following quote from you (with your name, if you agree) in the documentation of `nice_fit` so people know what they are in for? And also including the Jorgensen et al. (2018) reference in the corresponding section.

> I do not recommend including cutoffs in the table, as doing so would perpetuate their misuse. Fit indices are not test statistics, and their suggested cutoffs are not critical values associated with known Type I error rates. Numerous simulation studies have shown how poorly cutoffs perform in model selection (e.g., , [Jorgensen et al. (2018)](https://doi.org/10.1037/met0000152). Instead of test statistics, fit indices were designed to be measures of effect size (practical significance), which complement the chi-squared test of statistical significance. The range of RMSEA interpretations above is more reminiscent of the range of small/medium/large effect sizes proposed by Cohen for use in power analyses, which are as arbitrary as alpha levels, but at least they better respect the idea that (mis)fit is a matter of magnitude, not nearly so simple as "perfect or imperfect."

Second, I must admit that the cutoff guidance is a highly requested and appreciated feature, but at the same time, I am sensitive to the idea that such cutoff values should not be taken too seriously as demonstrated by simulation studies and so on (but does it have _any_ value for people who have not reference at all?). I like your idea of using "suggested cutoffs" instead of ideal values.

What about "Suggested _soft_ cutoffs" or simply "Commonly used cutoffs" to convey this idea even more strongly? Do you have a preference for any of the three?

Besides, I will correct the references and cutoffs and think about a way to include multiple sources.
