### 2023-10-07 93eedcf
add link to example 10 in vignette
