### 2023-10-07 rempsyc/lavaanExtra#21 @rempsyc
For reference, for the RMSEA (and other values), I relied on Table 3, with a column (rather boldly) called "**General rule for acceptable fit if data is continuous**":

![Fit indices recommendations](https://github.com/rempsyc/lavaanExtra/assets/13123390/87fac4f3-32b9-4ef0-82a3-2ee42a8cc6d4)

This is where I got the .05 [.00, .08] reference value. So I feel like it should be ok to cite this source as Table 3, and people can check the other references from that source if needed? Interestingly, Schreiber does not cite the references directly from the table either. I also like the idea of having a single reference source that people can look up for more information and references in the original literature.
