### 2023-10-07 5576e9a
closes #21 (cutoff values)
