### 2023-10-07 rempsyc/lavaanExtra#21 @rempsyc
OK, so for now I have:
1. Changed "Ideal value" for "Suggested soft cutoffs"
2. Specified that it is based on Schreiber (2017), Table 3, and not "proposed by".
3. Added an argument `cutoffs` so people can decide whether to include the reference row or not, based on their preference.

That said, I do feel like the word "cutoff" is more binary than ideal value, which in the way I was seeing it, was more expressive of the continuous nature of the fit indices while stile pointing the way toward more preferable values.
