### 2023-10-07 a50398a
closes #22 (add estimation of items reliability section in workflow vignette)
