+++ paper/paper.Rmd
@@ -120,9 +120,12 @@ library(tidySEM)
 ```{r}
 library(lavaanExtra)
 
-latent <- list(visual = paste0("x", 1:3),
-               textual = paste0("x", 4:6),
-               speed = paste0("x", 7:9))
+x <- paste0("x", 1:9)
+latent <- list(
+  visual = x[1:3],
+  textual = x[4:6],
+  speed = x[7:9]
+)
 
 model.cfa <- write_lavaan(latent = latent)
 cat(model.cfa)
@@ -150,7 +153,7 @@ IV <- c("grade", "ageyr")
 
 mediation <- list(speed = M, textual = M, visual = IV)
 regression <- list(speed = IV, textual = IV)
-covariance <- list(speed = "textual", ageyr = "grade", x4 = c("x5", "x6"))
+covariance <- list(speed = "textual", ageyr = "grade", x4 = x[5:6])
 indirect <- list(IV = IV, M = M, DV = DV)
 
 model.sem <- write_lavaan(mediation = mediation, 
