+++ README.Rmd
@@ -120,11 +120,12 @@ latent <- list(
   speed = c("x7", "x8", "x9")
 )
 
-# If you have many items, you can also use:
+# If you have many items, you can also use the `paste0` function:
+x <- paste0("x", 1:9)
 latent <- list(
-  visual = paste0("x", 1:3),
-  textual = paste0("x", 4:6),
-  speed = paste0("x", 7:9)
+  visual = x[1:3],
+  textual = x[4:6],
+  speed = x[7:9]
 )
 
 # Write the model, and check it
