+++ DESCRIPTION
@@ -18,7 +18,7 @@ Imports:
     lavaan,
     insight
 Suggests: 
-    rempsyc (> 0.1.2),
+    rempsyc (> 0.1.6),
     flextable,
     lavaanPlot,
     DiagrammeRsvg,
@@ -31,9 +31,11 @@ Suggests:
     tibble,
     sjlabelled,
     stringdist,
+    semTools,
+    psych,
+    testthat (>= 3.0.0),
     rmarkdown,
     markdown,
-    testthat (>= 3.0.0),
     covr,
     spelling
 Config/testthat/edition: 3
@@ -42,3 +44,4 @@ Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.3.9000
 VignetteBuilder: knitr
 Language: en-US
+Remotes: rempsyc/rempsyc
