+++ vignettes/fullworkflow.Rmd
@@ -105,9 +105,9 @@ unlink("fit_table.docx")
 Ideally, rather than just looking at the loadings, we would also estimate the item reliability of our dimensions for our long vs short scales to help select which items to drop for our short scale. We can first look at the alpha when an item is dropped using the `psych::alpha` function.
 
 ```{r, message=FALSE}
-visual <- HolzingerSwineford1939[names(HolzingerSwineford1939) %in% x[1:3]]
-textual <- HolzingerSwineford1939[names(HolzingerSwineford1939) %in% x[4:6]]
-speed <- HolzingerSwineford1939[names(HolzingerSwineford1939) %in% x[7:9]]
+visual <- HolzingerSwineford1939[x[1:3]]
+textual <- HolzingerSwineford1939[x[4:6]]
+speed <- HolzingerSwineford1939[x[7:9]]
 
 alpha(visual)
 alpha(textual)
