### 2023-10-07 rempsyc/lavaanExtra#22 @rempsyc
I like reporting omega, but also report alpha for "backward compatibility". I usually use `psych::omega`, and didn't know about `semTools::compRelSEM()`, I will look into it! (`psych::omega` also does not output the “Reliability if an item is dropped” section)

Oh, ok, I see, it is better for SEM (of course, coming from `semTools` :) ) because you can simply provide it your fit object instead of manually have to feed it the items. It fits great with the workflow in this case!

Using `psych::alpha()` fits not as well in the script because I have to give it the data manually. Here's what I've done so far, am I off-track?

https://lavaanextra.remi-theriault.com/articles/fullworkflow.html#estimation-of-items-reliability

---

I'm not sure if I've done this in the most efficient way. It seemed that I needed to repeat my code for each dimension, for both alpha and omega. If I had more time, my reflex would be to turn this `semTools::compRelSEM()` step into a function! But perhaps I should ask for a feature request in `semTools` instead :P
