### 2023-10-07 b22ddf7
closes #25 (process_vars and mediation argument)
