### 2023-10-07 rempsyc/lavaanExtra#25 @rempsyc
`process_vars()` is an internal function not useful for users, that is mainly useful for adding the + symbols, other operators, and section headers in a straightforward manner. It was not accessible before because I kept it within `write_lavaan()`. I have now pulled it out so the following will work:
```r
View(lavaanExtra:::process_vars)
```
However, I do not want to document it (using the `@noRd` tag), so `?process_vars()` will still not work, on purpose.

The `label` is indeed `FALSE` by default, and the differences are (1) the ability to name paths for mediations but not for regressions, and (2) the difference in section headers. I know it sounds a bit unnecessary, but I like to be able to accommodate different scenarios in the model specification output, so one section is clearly labeled regression, and the other clearly mediation, and so they can be used independently or jointly, etc. In this way people have (marginally) more freedom to express their model in the way they prefer.

So I do not think merging both arguments into one would be as satisfying to me. Maybe I misunderstood what you mean, if this is the case, please do not hesitate to clarify.
