### 2023-10-08 rempsyc/lavaanExtra#24 @rempsyc
> It would be wise to either (a) include a more thorough check for selecting only the rows of parameters that are indirect effects

Indeed, just like for total effects. I thought it made sense to lump them together in the mediation section of the model, but for `lavaan_ind()`, total effects and the moderated slope example you give (a good one) also come out with, which is inconsistent.

Is there any unique markers of indirect effects vs other user-defined parameters? I cannot find any, except perhaps the asterisk * which does not come up with the moderated slope example (but does for total effect, so not perfect). If not then perhaps I will have no choice but to change the function name and purpose to a more general one like you mention. Then I wonder what a good name would be.

Interestingly, `?model.syntax` does not include the `:=` operator in its list of seven operators (is it because it came later and the documentation was not updated, or perhaps it is not necessary in this section?). To be consistent with this family of functions, I am thinking something like `lavaan_user_defined_parameters()` but it is rather long, but shorter versions are not very clear: `lavaan_user_defined()`, `lavaan_user()`, `lavaan_parameters()`, `lavaan_custom()`, `lavaan_other()`...
