### 2023-10-08 dcd6f72
closes #24 (replaces lavaan_ind with lavaan_defined)
