### 2023-10-08 rempsyc/lavaanExtra#24 @rempsyc
`lavaan_ind()` is now named `lavaan_defined()`, and has the following new arguments (with their defaults):

```r
underscores_to_symbol = "\u2192",
lhs_name = "User-Defined Parameter",
rhs_name = "Paths",
```

So with your example we now get:

``` r
library(lavaan)
library(lavaanExtra)
fit <- sem(' x1 ~ c(a, b)*x2 ', data = HolzingerSwineford1939,
           group = "school", constraints = 'moderated_slope := a - b')
lavaan_defined(fit, underscores_to_symbol = "", nice_table = TRUE)
```
