### 2023-10-08 bcc72fe
#24 (accept a vector of replacements in lavaan_defined)
