### 2023-10-08 rempsyc/lavaanExtra#24 @rempsyc
> There could be an argument to [...] accept a vector of replacements if they are not all the same.

Done as well:

``` r
library(lavaan)
library(lavaanExtra)
latent <- list(visual = paste0("x", 1:3), textual = paste0("x", 4:6), speed = paste0("x", 7:9))
mediation <- list(speed = "visual", textual = "visual", visual = c("ageyr", "grade"))
indirect <- list(IV = c("ageyr", "grade"), M = "visual", DV = c("speed", "textual"))
HS.model <- write_lavaan(mediation, indirect = indirect, latent = latent, label = TRUE)
fit <- sem(HS.model, data = HolzingerSwineford1939)
lavaan_defined(fit, underscores_to_symbol = c("*", "+", "|", "~"), nice_table = TRUE)
```
