+++ R/lavaan_cor.R
@@ -15,10 +15,11 @@
 #'         variables, the correlation, and the corresponding p-value.
 #' @export
 #' @examplesIf requireNamespace("lavaan", quietly = TRUE)
+#' x <- paste0("x", 1:9)
 #' (latent <- list(
-#'   visual = paste0("x", 1:3),
-#'   textual = paste0("x", 4:6),
-#'   speed = paste0("x", 7:9)
+#'   visual = x[1:3],
+#'   textual = x[4:6],
+#'   speed = x[7:9]
 #' ))
 #'
 #' (regression <- list(
