+++ man/cfa_fit_plot.Rd
@@ -63,10 +63,11 @@ simultaneously.
 
 \examples{
 \dontshow{if (requireNamespace("lavaan", quietly = TRUE) && requireNamespace("lavaanPlot", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
+x <- paste0("x", 1:9)
 (latent <- list(
-  visual = paste0("x", 1:3),
-  textual = paste0("x", 4:6),
-  speed = paste0("x", 7:9)
+  visual = x[1:3],
+  textual = x[4:6],
+  speed = x[7:9]
 ))
 
 HS.model <- write_lavaan(latent = latent)
