+++ tests/testthat/test-nice_fit.R
@@ -28,8 +28,7 @@ test_that("nice_fit regular", {
 })
 
 test_that("nice_fit as nice_table", {
-  skip_on_os("linux")
-  skip_if_not_installed("rempsyc")
+   skip_if_not_installed("rempsyc")
   expect_s3_class(
     nice_fit(fit, nice_table = TRUE),
     c("nice_table", "flextable")
