### 2023-10-08 07e2afe
closes #18 (add lavaan_var)
