+++ tests/testthat/test-lavaan_cor.R
@@ -1,9 +1,10 @@
 suppressWarnings(library(lavaan))
 
+x <- paste0("x", 1:9)
 latent <- list(
-  visual = paste0("x", 1:3),
-  textual = paste0("x", 4:6),
-  speed = paste0("x", 7:9)
+  visual = x[1:3],
+  textual = x[4:6],
+  speed = x[7:9]
 )
 
 regression <- list(
