+++ R/nice_lavaanPlot.R
@@ -15,7 +15,6 @@
 #'  to .05). To plot all coefficients, set `sig` to 1.
 #' @param graph_options Read from left to right, rather than from top to bottom.
 #' @param ... Arguments to be passed to function [lavaanPlot::lavaanPlot].
-#' @keywords CFA lavaan plot fit
 #' @return A lavaanPlot, of classes `c("grViz", "htmlwidget")`, representing the
 #'         specified `lavaan` model.
 #' @export
