### 2023-10-08 rempsyc/lavaanExtra#18 @rempsyc
Thank you for these useful observations. I like these ideas and I think it makes the package stronger and more consistent. I have:
1. Deleted the `lavaan_cor()` file to avoid redundant code
2. Changed the `estimate` options (formerly b and B) to r and sigma in `lavaan_cov()`, with corresponding unicode support in `rempsyc::nice_table()`.
3. Kept only off-diagonal elements
4. Added in the same file `lavaan_cor()` as an alias to `lavaan_cov()`, but with a forced `estimate = "r"` for backward compatibility.
5. Added function `lavaan_var()` with only diagonal elements, with `estimate = sigma` by default, and the option to specify `estimate = r2` (in which case I provide `1 - sigma`, and corresponding corrected confidence intervals). I calculated the new r2 _p_ values like this:

```r
r2_pvalue <- function(est, se) {
  wald_z <- (1 - est) / se
  pnorm(wald_z)
}
```
So we get:

``` r
library(lavaan)
library(lavaanExtra)
x <- paste0("x", 1:9)
latent <- list(visual = x[1:3], textual = x[4:6], speed = x[7:9])
regression <- list(ageyr = c("visual", "textual", "speed"), grade = c("visual", "textual", "speed"))
covariance <- list(speed = "textual", ageyr = "grade")
HS.model <- write_lavaan(regression = regression, covariance = covariance, latent = latent, label = TRUE)
fit <- sem(HS.model, data = HolzingerSwineford1939)

lavaan_cov(fit, estimate = "sigma", nice_table = TRUE)
```

``` r
lavaan_cov(fit, estimate = "r", nice_table = TRUE)
```

``` r

lavaan_var(fit, estimate = "sigma", nice_table = TRUE)
```

``` r
lavaan_var(fit, estimate = "r2", nice_table = TRUE)
```

The fact that in this example I only get _p_ values of 1 is a bit suspicious, but it seems consistent with the previous _p_ values, which were all virtually 0, given that we have swapped the null hypothesis.

Finally,
> when estimate = "sigma" (with a \sigma^2 column header when nice_table=TRUE)

Wait, if it represents sigma squared, then should the estimation not be named `sigma2` instead of simply `sigma`?
