+++ paper/paper.Rmd
@@ -203,8 +203,9 @@ in tables like with the `nice_fit()` function.
 
 It is similarly possible to prepare APA tables in Word or other formats with the
 regression coefficients (`lavaan_reg()`), covariances (`lavaan_cov()`),
-correlations (`lavaan_cor()`), variances (`lavaan_var()`), or indirect effects 
-(`lavaan_defined()`). For example, for indirect effects:
+correlations (`lavaan_cor()`), variances (`lavaan_var()`), or user-defined 
+parameters like for indirect effects (`lavaan_defined()`). For example, for 
+indirect effects:
 
 ```{r, include=F, eval=F}
 x <- lavaan_defined(fit.sem, lhs_name = "Indirect Effect", nice_table = TRUE)
