### 2023-10-09 rempsyc/lavaanExtra#28 @rempsyc
For effect sizes in `lm` models, I like to report the semipartial correlation squared (delta R2), but for this I rely on `effectsize::r2_semipartial()`, which AFAIK does not work with `lavaan` models. In this case, I suppose you are correct that the standardized coefficient will need to suffice as the measure of effect size.

For any effect size, it is generally recommended to include the CI (and this seems consistent with Kelley & Preacher, 2012), so here I could report a single table with all required values, including the unstandardized estimate with its CI and _p_ value, AND the effect size (standardized estimate), also with its CI, but with no additional _p_ value.

I'm not sure about the `test` argument to report separate _p_ values with the 9 combinations... I wonder how niche this need is (and `lavaanExtra` isn't really supposed to be), so I wonder for those special needs whether it would be better to let people extract the info themselves with `standardizedsolution()`? I would tend to like to make the function both as simple and flexible as possible, but if you think it is important to accomodate this complex scenario, I will try to find a way to do it.

---

On another note, if I am not mistaken, this recommandation also applies to the new `lavaan_defined()` (formerly, `lavaan_ind()`), correct? Would this also apply to `lavaan_cov()` and `lavaan_var()` then?
