+++ R/lavaan_reg.R
@@ -6,9 +6,6 @@
 #'              See "Value" section for more details.
 #'
 #' @param fit lavaan fit object to extract fit indices from
-#' @param estimate What estimate to use, either the standardized
-#'                 estimate ("B", default), or unstandardized
-#'                 estimate ("b").
 #' @param nice_table Logical, whether to print the table as a
 #'                   [rempsyc::nice_table] as well as print the
 #'                   reference values at the bottom of the table.
@@ -37,21 +34,24 @@
 #' library(lavaan)
 #' fit <- sem(HS.model, data = HolzingerSwineford1939)
 #' lavaan_reg(fit)
-lavaan_reg <- function(fit, estimate = "B", nice_table = FALSE, ...) {
-  og.names <- c("lhs", "rhs", "pvalue", "est", "ci.lower", "ci.upper")
-  new.names <- c("Outcome", "Predictor", "p", "b", "CI_lower", "CI_upper")
-  if (estimate == "b") {
-    x <- lavaan::parameterEstimates(fit)
-  } else if (estimate == "B") {
-    x <- lavaan::standardizedsolution(fit, level = 0.95)
-    og.names[4] <- "est.std"
-    new.names[4] <- "B"
-  } else {
-    stop("The 'estimate' argument may only be one of c('B', 'b').")
-  }
+lavaan_reg <- function(fit, nice_table = FALSE, ...) {
+  og.names <- c("lhs", "rhs", "se", "z", "pvalue", "est", "ci.lower", "ci.upper")
+  new.names <- c("Outcome", "Predictor", "SE", "z", "p", "b", "CI_lower", "CI_upper", "B", "CI_lower_B", "CI_upper_B")
+
+  x <- lavaan::parameterEstimates(fit)
   x <- x[which(x["op"] == "~"), ]
   x <- x[og.names]
+
+  es <- lavaan::standardizedsolution(fit, level = 0.95)
+  es <- es[which(es["op"] == "~"), ]
+  es <- es[c("est.std", og.names[7:8])]
+
+  names(es)[2:3] <- paste0(names(es)[2:3], ".std")
+
+  x <- cbind(x, es)
+
   names(x) <- new.names
+
   if (nice_table) {
     insight::check_if_installed("rempsyc",
       version = get_dep_version("rempsyc"),
