### 2023-10-09 rempsyc/lavaanExtra#28 @rempsyc
So for now I have removed the estimate argument, so it is only possible to get one type of table. I also had to update `rempsyc::nice_table()` to support two confidence intervals in the same table... but it works :) reprex:

``` r
library(lavaan)
library(lavaanExtra)
x <- paste0("x", 1:9)
y <- c("visual", "textual", "speed")
latent <- list(visual = x[1:3], textual = x[4:6], speed = x[7:9])
regression <- list(ageyr = y, grade = y)
HS.model <- write_lavaan(regression = regression, latent = latent, label = TRUE)
fit <- sem(HS.model, data = HolzingerSwineford1939)

lavaan_reg(fit, nice_table = TRUE)
```

(tables appear ugly in reprexes but anyway)

With this, I hope we can move on with the paper but I am very interested to continue this discussion to bring the package up to the best reporting standards. An enormous thanks for all the time and expert advice you have put into improving this package. It is priceless.
