### 2023-10-09 #5701 @rempsyc
I would like to express my deepest gratitude for this extensive review that has greatly improved the package and brought it up to date with best practices. It has been tremendously helpful and it is not an understatement. Rarely are reviewers this generous of their time (I have not seen it at the JOSS/JOSE journals).

After a few days of work on these issues, I believe I have addressed most, if not all or them. For some of them, of course, the discussion will continue after the publication of this paper as we refine our perspectives on the best ways to present some information.

I am also sensitive to the idea of splitting miscellaneous packages into several specific ones, which is not a trivial decision. Like many, I started this package small, with only a few functions for my personal use, and it has grown over time to what it is now. I agree that it is early enough that a change of name, or a split into several packages, could be an advisable move for the long-term.

That said, I am not sure that I am willing to take such a decision before the imminent publication of the current paper. We are talking here about splitting the current package into three potential packages: (1) the alternative syntax (should we wish to keep it at all, though some find it useful), (2) publication-ready materials; and (3) automatic indirect effects (which actually relies on 1 currently to work). Because of the magnitude of such a change of direction, I believe it is better to delay this decision at this time.

The decision to delay the package splitting until after the publication of this paper may bring other problems in the future by further crystallizing the current reference, such as by slowing down the adopting of the new packages to be. At the same time, many have made similar at a later stage, so I think it will be possible in this case as well, given that there is sufficient perceived interest/need in what the package(s) will have to offer.

At this point, if all reviewers are satisfied with the current state of the paper (all boxes have been checked on the checklists), I would like to propose moving on with the next stage. With a bit of luck, I'll be able to include it in my fellowship application later this week :)
