### 2023-10-09 b2af2b4
correct lavaan_defined (#28)
