+++ tests/testthat/_snaps/lavaan_defined.md
@@ -3,16 +3,21 @@
     Code
       lavaan_defined(fit)
     Output
-         User.Defined.Parameter                       Paths     p      B CI_lower
-      30 ageyr → visual → speed   ageyr_visual*visual_speed 0.001 -0.151   -0.236
-      31 ageyr → visual → speed ageyr_visual*visual_textual 0.000 -0.153   -0.237
-      32 ageyr → visual → speed   grade_visual*visual_speed 0.000  0.248    0.150
-      33 ageyr → visual → speed grade_visual*visual_textual 0.000  0.252    0.160
-         CI_upper
-      30   -0.066
-      31   -0.070
-      32    0.345
-      33    0.344
+         User-Defined Parameter                       Paths         SE         Z
+      30 ageyr → visual → speed   ageyr_visual*visual_speed 0.02808889 -3.198387
+      31 ageyr → visual → speed ageyr_visual*visual_textual 0.04191650 -3.461890
+      32 ageyr → visual → speed   grade_visual*visual_speed 0.07291514  4.257496
+      33 ageyr → visual → speed grade_visual*visual_textual 0.10134908  4.947490
+                    p           b   CI_lower    CI_upper          B CI_lower_B
+      30 1.381987e-03 -0.08983914 -0.1448924 -0.03478593 -0.1508037 -0.2358595
+      31 5.363956e-04 -0.14511033 -0.2272652 -0.06295550 -0.1534909 -0.2371048
+      32 2.067294e-05  0.31043593  0.1675249  0.45334698  0.2477787  0.1503789
+      33 7.517664e-07  0.50142352  0.3027830  0.70006406  0.2521937  0.1601663
+          CI_upper_B
+      30 -0.06574796
+      31 -0.06987694
+      32  0.34517843
+      33  0.34422119
 
 # nice_fit as nice_table
 
@@ -20,61 +25,32 @@
       lavaan_defined(fit, nice_table = TRUE)
     Output
       a flextable object.
-      col_keys: `User-Defined Parameter`, `Paths`, `p`, `B`, `95% CI` 
+      col_keys: `User-Defined Parameter`, `Paths`, `SE`, `Z`, `p`, `b`, `95% CI (b)`, `B`, `95% CI (B)` 
       header has 1 row(s) 
       body has 4 row(s) 
       original dataset sample: 
-         User-Defined Parameter                       Paths            p          B
-      30 ageyr → visual → speed   ageyr_visual*visual_speed 5.108293e-04 -0.1508037
-      31 ageyr → visual → speed ageyr_visual*visual_textual 3.207669e-04 -0.1534909
-      32 ageyr → visual → speed   grade_visual*visual_speed 6.163554e-07  0.2477787
-      33 ageyr → visual → speed grade_visual*visual_textual 7.824837e-08  0.2521937
-                 95% CI
-      30 [-0.24, -0.07]
-      31 [-0.24, -0.07]
-      32   [0.15, 0.35]
-      33   [0.16, 0.34]
-
-# nice_fit estimates
-
-    Code
-      lavaan_defined(fit, estimate = "b")
-    Output
-         User.Defined.Parameter                       Paths     p      b CI_lower
-      30 ageyr → visual → speed   ageyr_visual*visual_speed 0.001 -0.090   -0.145
-      31 ageyr → visual → speed ageyr_visual*visual_textual 0.001 -0.145   -0.227
-      32 ageyr → visual → speed   grade_visual*visual_speed 0.000  0.310    0.168
-      33 ageyr → visual → speed grade_visual*visual_textual 0.000  0.501    0.303
-         CI_upper
-      30   -0.035
-      31   -0.063
-      32    0.453
-      33    0.700
-
----
-
-    Code
[…]
