### 2023-10-09 f633db1
* new function: `lavaan_extract()`, which takes a specific operator and extracts relevant information (now used internally for `lavaan_reg()`, `lavaan_defined()`)
