+++ R/lavaan_extract.R
@@ -1,0 +1,99 @@
+#' @title Extract relevant indices from lavaan model based on specified operator
+#'
+#' @description Extract relevant indices from lavaan model through
+#'  [lavaan::parameterEstimates] and [lavaan::standardizedsolution].
+#' @param fit lavaan fit object to extract fit indices from
+#' @param operator Which operator to subselect with.
+#' @param lhs_name Name of first column, referring to the left-hand side
+#'  expression (lhs).
+#' @param rhs_name Name of first column, referring to the right-hand side
+#'  expression (rhs).
+#' @param underscores_to_symbol Character to convert underscores
+#'  to arrows in the first column, like for indirect effects. Default to
+#'  the arrow symbol (→), but can be set to NULL or "_", or to any
+#'  other desired  symbol. It is also possible to provide a vector of
+#'  replacements if they they are not all the same.
+#' @param nice_table Logical, whether to print the table as a
+#'                   [rempsyc::nice_table] as well as print the
+#'                   reference values at the bottom of the table.
+#' @param ... Arguments to be passed to [rempsyc::nice_table]
+#' @return A dataframe, including the indirect effect ("lhs"),
+#'         corresponding paths ("rhs"), standardized regression
+#'         coefficient ("std.all"), corresponding p-value, as well
+#'         as the unstandardized regression coefficient ("est") and
+#'         its confidence interval ("ci.lower", "ci.upper").
+#' @export
+#' @examplesIf requireNamespace("lavaan", quietly = TRUE)
+#' x <- paste0("x", 1:9)
+#' (latent <- list(
+#'   visual = x[1:3],
+#'   textual = x[4:6],
+#'   speed = x[7:9]
+#' ))
+#'
+#' (mediation <- list(
+#'   speed = "visual",
+#'   textual = "visual",
+#'   visual = c("ageyr", "grade")
+#' ))
+#'
+#' (indirect <- list(
+#'   IV = c("ageyr", "grade"),
+#'   M = "visual",
+#'   DV = c("speed", "textual")
+#' ))
+#'
+#' HS.model <- write_lavaan(mediation,
+#'   indirect = indirect,
+#'   latent = latent, label = TRUE
+#' )
+#' cat(HS.model)
+#'
+#' library(lavaan)
+#' fit <- sem(HS.model, data = HolzingerSwineford1939)
+#' lavaan_extract(fit, lhs_name = "Indirect Effect", operator = ":=")
+lavaan_extract <- function(fit,
+                           operator = NULL,
+                           lhs_name = "Left-Hand Side",
+                           rhs_name = "Right-Hand Side",
+                           underscores_to_symbol = "\u2192",
+                           nice_table = FALSE,
+                           ...) {
+  if (missing(operator)) {
+    stop("Please specify the desired operator.")
+  }
+  og.names <- c("lhs", "rhs", "se", "z", "pvalue", "est", "ci.lower", "ci.upper")
+  new.names <- c(lhs_name, rhs_name, "SE", "Z", "p", "b",
+                 "CI_lower", "CI_upper", "B", "CI_lower_B", "CI_upper_B")
+
+  x <- lavaan::parameterEstimates(fit)
+  x <- x[which(x["op"] == operator), ]
+  x <- x[og.names]
+
+  es <- lavaan::standardizedsolution(fit, level = 0.95)
+  es <- es[which(es["op"] == operator), ]
+  es <- es[c("est.std", og.names[7:8])]
+  names(es)[2:3] <- paste0(names(es)[2:3], ".std")
+
+  x <- cbind(x, es)
+
+  names(x) <- new.names
+
[…]
