+++ R/lavaan_defined.R
@@ -3,13 +3,8 @@
 #'
 #' @description Extract relevant user-defined parameters (e.g., indirect or
 #'  total effects) indices from lavaan model through
-#'  [lavaan::parameterEstimates] with `standardized = TRUE`. In this case,
-#'  the beta (B) represents the resulting `std.all` column. See "Value" section
-#'  for more details.
+#'  [lavaan::parameterEstimates] and [lavaan::standardizedsolution].
 #' @param fit lavaan fit object to extract fit indices from
-#' @param nice_table Logical, whether to print the table as a
-#'                   [rempsyc::nice_table] as well as print the
-#'                   reference values at the bottom of the table.
 #' @param underscores_to_symbol Character to convert underscores
 #'  to arrows in the first column, like for indirect effects. Default to
 #'  the arrow symbol (→), but can be set to NULL or "_", or to any
@@ -19,6 +14,9 @@
 #'  expression (lhs).
 #' @param rhs_name Name of first column, referring to the right-hand side
 #'  expression (rhs).
+#' @param nice_table Logical, whether to print the table as a
+#'                   [rempsyc::nice_table] as well as print the
+#'                   reference values at the bottom of the table.
 #' @param ... Arguments to be passed to [rempsyc::nice_table]
 #' @return A dataframe, including the indirect effect ("lhs"),
 #'         corresponding paths ("rhs"), standardized regression
@@ -57,45 +55,17 @@
 #' fit <- sem(HS.model, data = HolzingerSwineford1939)
 #' lavaan_defined(fit, lhs_name = "Indirect Effect")
 lavaan_defined <- function(fit,
-                           nice_table = FALSE,
                            underscores_to_symbol = "\u2192",
                            lhs_name = "User-Defined Parameter",
                            rhs_name = "Paths",
+                           nice_table = FALSE,
                            ...) {
-  og.names <- c("lhs", "rhs", "se", "z", "pvalue", "est", "ci.lower", "ci.upper")
-  new.names <- c(lhs_name, rhs_name, "SE", "Z", "p", "b", "CI_lower", "CI_upper", "B", "CI_lower_B", "CI_upper_B")
-
-  x <- lavaan::parameterEstimates(fit)
-  x <- x[which(x["op"] == ":="), ]
-  x <- x[og.names]
-
-  es <- lavaan::standardizedsolution(fit, level = 0.95)
-  es <- es[which(es["op"] == ":="), ]
-  es <- es[c("est.std", og.names[7:8])]
-
-  names(es)[2:3] <- paste0(names(es)[2:3], ".std")
-
-  x <- cbind(x, es)
-
-  names(x) <- new.names
-
-  if (!is.null(underscores_to_symbol)) {
-    if (length(underscores_to_symbol) == 1 || length(underscores_to_symbol) == nrow(x)) {
-      x[[1]] <- unlist(lapply(seq_along(underscores_to_symbol), function(i) {
-        gsub("_", paste0(" ", underscores_to_symbol[[i]], " "), as.list(x[[1]])[[i]])
-      }))
-    } else {
-      stop("'underscores_to_symbol' must match the number of rows.")
-    }
-  }
-  if (nice_table) {
-    insight::check_if_installed("rempsyc",
-      version = get_dep_version("rempsyc"),
-      reason = "for this feature."
-    )
-    x <- rempsyc::nice_table(x, ...)
-  }
-  x
+  lavaan_extract(fit,
+                 operator = ":=",
+                 lhs_name = lhs_name,
+                 rhs_name = rhs_name,
+                 underscores_to_symbol = underscores_to_symbol,
+                 nice_table = nice_table)
 }
 
 lavaan_ind <- lavaan_defined
