### 2023-10-09 rempsyc/lavaanExtra#28 @rempsyc
Both of these also now rely on a new internal function, [`lavaan_extract()`](https://lavaanextra.remi-theriault.com/reference/lavaan_extract.html), which I am hoping to integrate with `lavaan_var()` and `lavaan_cov()` as well.
