### 2023-10-09 8052819
use `lavaan_extract()` internally for `lavaan_cov()`)
