+++ R/lavaan_extract.R
@@ -13,6 +13,8 @@
 #'  the arrow symbol (→), but can be set to NULL or "_", or to any
 #'  other desired  symbol. It is also possible to provide a vector of
 #'  replacements if they they are not all the same.
+#' @param diag When extracting covariances (`~~`), whether to include or
+#'  exclude diagonal values (one of "exclude" or "include").
 #' @param nice_table Logical, whether to print the table as a
 #'                   [rempsyc::nice_table] as well as print the
 #'                   reference values at the bottom of the table.
