### 2023-10-09 rempsyc/lavaanExtra#21 @TDJorgensen
> I also like the idea of having a single reference source that people can look up for more information and references in the original literature.

The problem is that Schreiber was doing the same thing you are doing (providing a quick summary of many different original sources), but only for a particular field. And he didn't seem to summarize accurately or with adequate references to original sources.  So I wouldn't rely on this as your citation.

> I feel like it should be ok to cite this source as Table 3, and people can check the other references from that source if needed?

I don't even think it is necessary to provide a citation in the table, because writers should be discussing their own decision rules in their actual text, where citations can be more thorough than in a table note.  My recommendation for you is to include more detail in your `nice_fit()` documentation, where there is room for more thorough citation of original sources.  You could have a bullet-point list in the **Details** section, where you list and cite the full range of interpretations for RMSEA, the .90 and .95 guidelines for incremental fit indices, etc.

> I do feel like the word "cutoff" is more binary than ideal value, which ... was more expressive of the continuous nature of the fit indices

I agree, but the "ideal" value is 1 for goodness-of-fit indices (CFI and TLI) and 0 for badness-of-fit indices (RMSEA and SRMR).  How about "Common guidelines" for the bottom row?  And it can include alternatives (e.g., "> .90 or .95" for CFI), but let the readers find the details in your documentation.  If you want to be sure to draw their attention to it, then the `nice_fit()` function could print a `message()` when both `cutoffs=TRUE` and `nice_table=TRUE`, e.g.,
```
if (cutoffs && nice_table) {
  message("See the ?nice_fit help-page Details for more information about guidelines")
}
```
