### 2023-10-09 rempsyc/lavaanExtra#22 @TDJorgensen
The example definitely gets more complex due to modeling more than 1 dimension.  I don't think it is necessary to include omega, since the focus is just to illustrate when (and why) the `remove.items=` argument can be useful. The `alpha()` examples serve as a quick hypothetical justification, which is all you need.
