### 2023-10-09 rempsyc/lavaanExtra#18 @TDJorgensen
> in this example I only get p values of 1

You are calculating the wrong `wald_z` statistic, as well as the wrong *p* value.

```
wald_z <- (1 - est) / se
```
At this point in the syntax, `est` is already R^2, so there is no reason to subtract it from 1 again to make it the proportion of *un*explained variance.  If you did, you would have to additionally subtract the null hypothesized value of 1 (i.e., no variance is explained, so 100% is unexplained). That would make it  `1 - est - 1` which is simply `-est`.   And you would have to calculate the _p_ value as the `lower.tail=TRUE` of the curve (which is what you are currently doing.

It is simpler to calculate `wald_z <- est/se` for the null hypothesis that there is no explained variance, but the _p_ value should be the upper tail of the curve: `pnorm(wald_z, lower.tail = FALSE)`.

> if it represents sigma squared, then should the estimation not be named sigma2 instead of simply sigma?

The table column should certainly be labeled \sigma^2 because it is a variance estimate, not a _SD_ estimate.  Maybe you are right that the argument should be `"sigma2"` then, which also matches the "squared" in `"r2"`.

For `lavaan_var()` it is not necessary to have 2 identical columns of variable names.  You can drop the `rhs`.
