### 2023-10-09 rempsyc/lavaanExtra#28 @TDJorgensen
> this recommandation also applies to the new lavaan_defined() (formerly, lavaan_ind()), correct?

Yes, nice work.  I think these tables include all the right information, and no need to accommodate complicated hypothetical combinations that I speculated about.

> Would this also apply to lavaan_cov() and lavaan_var() then?

No, I don't think people typically have *a priori* hypotheses they want to test about most of those parameters anyway.  Except for factor correlations, whose metric is arbitrary anyway.

> For effect sizes in `lm` models, I like to report the semipartial correlation squared (delta R2), but for this I rely on `effectsize::r2_semipartial()`, which AFAIK does not work with `lavaan` models

I doubt it, but this paper could be helpful if you want to put it on your long-term feature-request list:

https://doi.org/10.3758/s13428-020-01532-y
