### 2023-10-09 #5701 @TDJorgensen
I'm fine with the current states of the paper and package, except for a mistake I found in [this issue](https://github.com/rempsyc/lavaanExtra/issues/18) (that was already closed).
