### 2023-10-09 rempsyc/lavaanExtra#18 @rempsyc
Ah, yes, I see, thanks for the explanations! So correction is:

```r
x$est.std <- abs(1 - x$est.std)
x$pvalue <- stats::pnorm(x$est.std / x$se, lower.tail = FALSE)
```

And we get:

``` r
library(lavaan)
library(lavaanExtra)
x <- paste0("x", 1:9)
latent <- list(visual = x[1:3], textual = x[4:6], speed = x[7:9])
mediation <- list(speed = "visual", textual = "visual", visual = c("ageyr", "grade"))
indirect <- list(IV = c("ageyr", "grade"), M = "visual", DV = c("speed", "textual"))
HS.model <- write_lavaan(mediation, indirect = indirect, latent = latent, label = TRUE)
fit <- sem(HS.model, data = HolzingerSwineford1939)

lavaan_var(fit, estimate = "r2", nice_table = TRUE)
```

``` r
lavaan_var(fit, estimate = "sigma2", nice_table = TRUE)
```

And now things look better :)
