### 2023-10-09 rempsyc/lavaanExtra#28 @rempsyc
Well, I tried integrating `lavaan_cov()` with `lavaan_extract()`, so we get a consistent structure always with both standardized and unstandardized estimates, and simplified code internally. So again here it is not possible to specify `estimate = r` or `estimate = sigma` since you will always get a single table with both.

``` r
library(lavaan)
library(lavaanExtra)
x <- paste0("x", 1:9)
latent <- list(visual = x[1:3], textual = x[4:6], speed = x[7:9])
covariance <- list(speed = "textual", ageyr = "grade")
HS.model <- write_lavaan(covariance = covariance, latent = latent, label = TRUE)
fit <- sem(HS.model, data = HolzingerSwineford1939)

lavaan_cov(fit, nice_table = TRUE)
```

The only issue is that now I only report _p_ values for the sigma, so I wonder if people may misreport the correlations with the sigma _p_ values in this case (though it's a bit unclear to me how this scenario is different from looking at the standardized coefficient, _r_, as simply the standardized effect size as with the other extracting functions).
