### 2023-10-09 5e1144a
remove omega complexity from example (#22)
