+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: lavaanExtra
 Title: Convenience Functions for Package 'lavaan'
-Version: 0.1.8.1
-Date: 2023-08-25
+Version: 0.1.9
+Date: 2023-10-09
 Authors@R: 
     person("Rémi", "Thériault", , "remi.theriault@mail.mcgill.ca", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-4315-6788"))
@@ -43,4 +43,3 @@ Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.3.9000
 VignetteBuilder: knitr
 Language: en-US
-Remotes: rempsyc/rempsyc
