### 2023-10-09 #5701 @rempsyc
Great, thank you!

Here is the release (version 0.1.9): https://github.com/rempsyc/lavaanExtra/releases/tag/v0.1.9

Here is the Zenodo archive and doi: https://doi.org/10.5281/zenodo.8421873
