### 2023-10-09 #5701 @arfon
@rempsyc – please take a look at this final draft PDF. I note there's a fair number of things that look like `{lavaanExtra}` in the text which I'm assuming is a formatting mistake?
