+++ paper/paper.bib
@@ -26,12 +26,15 @@
     url = {https://CRAN.R-project.org/package=flextable},
   }
 
-@Misc{rempsycPackage,
+@Article{rempsycPackage,
     title = {{rempsyc}: Convenience Functions for Psychology},
     author = {Rémi Thériault},
-    year = {2022},
-    note = {(R package version 0.1.1.6) [Computer software].},
-    url = {https://rempsyc.remi-theriault.com},
+    journal = {Journal of Open Source Software},
+    year = {2023},
+    volume = {8},
+    number = {87},
+    pages = {5466},
+    doi = {10.21105/joss.05466},
   }
 
 @Manual{lavaanPlotPackage,
