### 2023-10-09 #5701 @rempsyc
Originally, I used code formatting for packages, like `lavaanExtra`, but my colleagues from easystats have suggested that it is better for JOSS to actually use brackets {lavaanExtra} when referring to packages ([a recently published example](https://doi.org/10.21105/joss.04684)). So, what is the official preference? I can change it based on that.

Note: I have added to doi for `rempsyc` since it has now been published in JOSS as well.
