### 2023-10-11 #5701 @arfon
> So, what is the official preference? I can change it based on that.

I don't think we have one. I personally prefer `lavaanExtra` (hence my comment) but I'll leave it up to you!
