+++ paper/paper.md
@@ -13,7 +13,7 @@ authors:
 affiliations:
   - name: "Department of Psychology, Université du Québec à Montréal, Québec, Canada"
     index: 1
-date: "2023-10-09"
+date: "2023-10-11"
 bibliography: paper.bib
 output:
   rticles::joss_article:
@@ -31,28 +31,28 @@ editor_options:
 
 # Summary
 
-{lavaanExtra} is an R package that offers an alternative and vector-based
-syntax to the package {lavaan}, as well as other convenience functions such
+`lavaanExtra` is an R package that offers an alternative and vector-based
+syntax to the package `lavaan`, as well as other convenience functions such
 as naming paths and defining indirect effects automatically. It also
 offers convenience formatting optimized for publication and script
 sharing workflows.
 
 # Statement of need
 
-{lavaan} [@lavaanPackage] is a very popular R package for structural
+`lavaan` [@lavaanPackage] is a very popular R package for structural
 equation modeling (SEM). The package relies on specific operators to define
 latent variables, regressions, covariances, indirect effects, and so on.
-However, some individuals (e.g., beginners to R and {lavaan})—or in some
+However, some individuals (e.g., beginners to R and `lavaan`)—or in some
 cases power users—may prefer not having to specify the operators themselves, 
-or would like to see some steps automatized, such as generating the {lavaan} 
+or would like to see some steps automatized, such as generating the `lavaan` 
 model layout or defining indirect effects. Furthermore, for researchers, it 
 can be relatively difficult to extract relevant statistical outputs in the 
 form of tables and figures that are suitable for scientific publication.
 
-{lavaanExtra} does mainly two things to address these issues. First, it offers 
+`lavaanExtra` does mainly two things to address these issues. First, it offers 
 an alternative, code-efficient flexible modular syntax that allows automatizing 
 certain steps, such as defining indirect effects in certain scenarios or the 
-desired structure of a SEM model to be plotted (however, note that {lavaan} is
+desired structure of a SEM model to be plotted (however, note that `lavaan` is
 also compatible with a modular approach). Second, it facilitates the 
 analysis-to-publication workflow by providing publication-ready tables and 
 figures following the style requirements of the American Psychological 
@@ -69,7 +69,7 @@ specific symbols required for each aspect of the model.
 
 There are several benefits to this approach. Some `lavaan` models can
 become very large. By defining the entire model every time, such as is
-typical with {lavaan} users, not only do we break the DRY (Don't Repeat
+typical with `lavaan` users, not only do we break the DRY (Don't Repeat
 Yourself) principle, but our scripts can also become long and unwieldy.
 This problem gets worse in the scenario where we want to compare several
 variations of the same general model. `write_lavaan()` allows the user
@@ -86,7 +86,7 @@ The vector-based approach also allows the use of functions to define
 components. For example, if all scale items are named consistently, say
 `x1` to `x50`, one can use `paste0("x", 1:50)` instead of typing all the
 items by hand and risk making mistakes. However, note that reusable components
-through functions is also compatible with {lavaan}.
+through functions is also compatible with `lavaan`.
 
 Another issue with `lavaan` models is the readability of the code defining
 the model. One can go to lengths to make it pretty, but not everyone
@@ -145,7 +145,7 @@ predicted by our mediating variable, `visual`. In turn, `visual` is now
 predicted by our independent variables, `grade` (the students' grade) and 
 `ageyr` (the students' age, in years).
 
[…]
